requirements.txt
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mysql-connector-python==8.0.32
packaging==23.0
protobuf==3.20.3
python-engineio==4.3.4
python-socketio==5.7.2
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
uuid==1.30
Werkzeug==2.2.3
wsproto==1.2.0
